chore(pi-authentik): restructure extension folder layout
