+++ extensions/pi-authentik/docs/LLM_ENDPOINT_SETUP.md
