+++ extensions/pi-authentik/src/auth/discovery.ts
