+++ extensions/pi-authentik/src/auth/logout.ts
