+++ extensions/pi-authentik/src/llm/endpoint-validator.ts
