+++ extensions/pi-authentik/src/llm/models.ts
