+++ extensions/pi-authentik/src/tests/readme.test.ts
@@ -6,7 +6,10 @@ import test from "node:test";
 const extensionDir = path.resolve(import.meta.dirname);
 
 function readDoc(name: string): string {
-  return fs.readFileSync(path.join(extensionDir, name), "utf8");
+  const relPath = name === "README.md" || name === ".env.example"
+    ? path.join("../..", name)
+    : path.join("../../docs", name);
+  return fs.readFileSync(path.resolve(extensionDir, relPath), "utf8");
 }
 
 test("README covers setup, commands, /v1, troubleshooting, pi-secret storage, and settings-based config", () => {
@@ -48,7 +51,7 @@ test("AUTHENTIK_SETUP explains that the regex is anchored and restricts to 127.0
   // Doc must explain the anchoring and host restriction
   assert.match(setup, /anchored/i);
   assert.match(setup, /127\.0\.0\.1.*only/i);
-  assert.match(setup, /localhost.*(not|excluded|is not allowed|is excluded)/i);
+  assert.match(setup, /(not|excluded|is not allowed|is excluded).*localhost/i);
   assert.match(setup, /one or more digits/i); // Doc describes \d+ for port
 });
